Squish warning

+++ fbink.c
@@ -108,7 +108,7 @@ static void
     rotate_coordinates(FBInkCoordinates* coords)
 {
 	unsigned short int rx = coords->y;
-	unsigned short int ry = viewWidth - coords->x - 1;
+	unsigned short int ry = (unsigned short int) (viewWidth - coords->x - 1);
 
 	coords->x = rx;
 	coords->y = ry;
